add index ok

+++ jcart-administration-back/src/main/java/io/bnn/jcartadministrationback/dao/ReturnHistoryMapper.java
@@ -19,5 +19,5 @@ public interface ReturnHistoryMapper {
 
     int updateByPrimaryKey(ReturnHistory record);
 
-    List<ReturnHistory> getListByReturnId(Integer returnId);
+    List<ReturnHistory> getListByReturnId(Long returnId);
 }
