+++ jcart-administration-back/src/main/java/io/bnn/jcartadministrationback/service/ReturnHistoryService.java
@@ -5,7 +5,7 @@ import io.bnn.jcartadministrationback.po.ReturnHistory;
 import java.util.List;
 
 public interface ReturnHistoryService {
-    List<ReturnHistory> getListByReturnId(Integer returnId);
+    List<ReturnHistory> getListByReturnId(Long returnId);
 
     Long create(ReturnHistory returnHistory);
 }
